refactor: move User schema and model into models directory
